Start getting things ready for server-side rendering
Add Browser.server.js kernel code. This will replace Browser.js when
generating code to run on the server.

Replace Browser.Url with a String. This way the code to translate a URL
into nice structured data can live in Elm and will definitely work the
same when on client and server.

Stop giving the module name on initialization. I don’t think it is
necessary anymore.