Address issue #16, "Debugger click-to-resume overlay broken on 0.19"
Add the "elm-debugger-overlay" id to the overlay so that the event listener in Debugger.js, (in the _Debugger_blocker function), can recognize the click event on the overlay and let it through so the program can resume.

Add some styling so the overlay displays as expected.

Note, there are still some issues with multiple popouts getting spawned, but that's for a different day.