Allow debugger to use iframe browsing-context
This change allows an existing iframe with the attribute "name" being "elm-debugger" to be used as the browsing context for the debugger. This way, the application or page can choose to have the debugger integrated instead of having it outside in another window. Some of developers (me) are unable to focus well enough to not loose track of a debug window and it's associated page.